More fixes & improvements
General
* More time counters.
* Add callable_ref, replacing std::function.
* Fix wrap_angle.

Graphics
* Add debug auto-screenshot.
* Add shield regeneration effect.
* Fix screen fade effect.

Gameplay
* Add raycast with width.
* Fix shields - no negative HP; no healing under damage
when disabled; damage absorption on destuction is limited.
* Tweak player and drone shield params.

Weapons
* Use integer ammo.
* Add optional friendly fire.
* Fix shooting when facing a wall.
* Fix direct hit with AoE projectile.
* Tweak params on all weapons.
* Minigun: variable dispersion, alt fire.
* Bolter: main fire ray has width.
* Foamgun: has ammo, projectiles improved.
* Rifle: grenades improved.

+++ game/s_objs.hpp
@@ -47,7 +47,7 @@ public:
 	struct AmmoPack
 	{
 		AmmoType type;
-		float amount;
+		int amount;
 	};
 	
 	struct Func
